requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.2
pillow==11.3.0
polars==1.33.1
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
sympy==1.14.0
torch==2.2.1
torchvision==0.17.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.2.35
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.35.0
